Preprocessor enhancement (#27)
* feat: Introduce additonal preprocessor functionality to compiler

Added preprocessor support to the compiler project, enabling macro definitions, conditional compilation, and predefined macros. Included `preprocessor.cpp` and `preprocessor.h` in the project and updated `parser.cpp` and `toker.cpp` to utilize the new preprocessor features.

* fix: Ignore any whitespaces with preprocessor directives

Doing this allows for whitespaces to be ignored/stripped when a preprocessor declaritive has been found. This fix is to allow for custom formatting of blitzbasic code.

* Optimized preprocessor

* Localization

---------

Co-authored-by: 子悦解说 <ziyuecommentary@gmail.com>